Adding Story Pane Sub Menu (1/2)
Committed by mrkriegh: Changed a few things to get this working. First, I moved the main menu items dictionary to the menu script so that it could designate calls to the menu functions. Before, with the dictionary in the main script, I was running into interpreting order errors. Next, I created the story pane and gave it some place holder text to be able to display something. Finally, I screwed around with the silly menu control cooldowns until I was able to not risk epileptic users. Now, the submenu correctly shows up when you select it, and disappears when you press SPACEBAR while it is displayed. Next commit will be correctly positioning the story pane and giving it something of a finalized story text.