Adding Story Pane Sub Menu (2/2)
Committed by mrkriegh: Added a more finalized version of the story. I may go back and fix it later, but it is good enough for done for this project. I also adjusted placement and spacing of the text pane and the text. Looks good enough for now. Not perfect, but good enough.