Basic Audio Menu
Committed by mrkriegh: Displaying a basic audio menu. Corrected the 'back to previous menu' button to be ESCAPE, since the upgrade menu uses the SPACE button to interact with the currently displayed menu. I then utilized a lot of the same code that is used in upgrade.py to generate the audio menu. I was planning on rotating the orientation of the audio menu, but I decided to put it in the same style as the upgrade menu to keep things a little more uniform. I may have to go back and fix this for the main menu at some point. As of yet, you cannot change the volume values, but that will be covered in issue #2.