Player Death
Committed by mrkriegh: Added a new flag and a couple of functions to the Level script so that I can track if a player is dead and react accordingly. Meaning I wait a brief period of time and kick the user back to the main menu. To keep things simple and running fast, I gave the player the Spirit's death animation and the generic death audio. Now when the user is reduced to zero health, their character will explode, the game will wait briefly, then return the user to the main menu. I think I might add the player's score to the debug message at that point, and maybe any future communication with the player, as it is quite handy.