Adding Score Controller
Committed by mrkriegh: Added a Score Controller to be able to track how many kills the player makes and display their high score on the main menu. Rearranged the level code to initialize the flags and map through an initialize method that gets called every time the user starts a new game, not just when the level code object is created. This allows for the level to be reset every time the user starts a new game without having to quit the program in between. Also destroyed all enemies and the player at every game over point to help with this clean up process. I then implemented two game over points, when the player is defeated, and when the player defeats every monster on the map. The second option is really only for the tutorial level. Finally, I cleaned up some debug messages so that I could display the score during play, the high score at the main menu, and the Game Over Message at the game over points.